+++ client/package.json
@@ -12,6 +12,7 @@
     "react-announcement-bar": "^0.0.1",
     "react-chartjs-2": "^2.7.4",
     "react-dom": "^16.4.2",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
     "semantic-ui-css": "^2.3.3",
     "semantic-ui-react": "^0.82.2"
